chore: move deployment & startInstance to app
