+++ client/src/plugins/zeebe-plugin/start-instance-plugin/__tests__/StartInstancePluginSpec.js
@@ -19,7 +19,7 @@ import { mount } from 'enzyme';
 import StartInstancePlugin from '../StartInstancePlugin';
 
 import { Slot, SlotFillRoot } from '../../../../app/slot-fill';
-import { DEFAULT_ENDPOINT } from '../../../../remote/Deployment';
+import { DEFAULT_ENDPOINT } from '../../../../app/zeebe/Deployment';
 
 import { Deployment, StartInstance, ZeebeAPI } from '../../../../app/__tests__/mocks';
